+++ packages/react-tinacms-editor/src/components/ProsemirrorEditor/index.tsx
@@ -19,11 +19,13 @@ limitations under the License.
 import * as React from 'react'
 import { useEffect, useRef, useState } from 'react'
 import styled from 'styled-components'
+import { EditorView } from 'prosemirror-view'
 
 import { EditorProps } from '../../types'
 import { EditorStateProvider } from '../../context/editorState'
-import { Menubar } from './Menubar'
+import { buildEditor } from '../../context/utils/buildEditor'
 
+import { Menubar } from './Menubar'
 import { CodeMirrorCss } from './styles/CodeMirror'
 import { ProseMirrorCss } from './styles/ProseMirror'
 
@@ -36,12 +38,18 @@ export const ProsemirrorEditor = styled(
     imageProps,
     ...styleProps
   }: EditorProps) => {
-    const [el, setEl] = useState<HTMLDivElement>()
     const editorRef = useRef<HTMLDivElement | null>(null)
+    const [editorView, setEditorView] = useState<{ view: EditorView }>()
+    const [translator, setTranslator] = useState<any>()
 
     useEffect(() => {
       // State is updated with latest value of editorRef to trigger re-render
-      if (editorRef.current) setEl(editorRef.current)
+      setTranslator(
+        buildEditor(input, editorRef.current, imageProps, setEditorView, format)
+      )
+      return () => {
+        editorView && editorView.view.destroy()
+      }
     }, [editorRef])
 
     return (
@@ -50,12 +58,7 @@ export const ProsemirrorEditor = styled(
           rel="stylesheet"
           href="https://codemirror.net/lib/codemirror.css"
         />
-        <EditorStateProvider
-          input={input}
-          el={el}
-          imageProps={imageProps}
-          format={format}
-        >
+        <EditorStateProvider translator={translator} editorView={editorView}>
           <Menubar
             sticky={sticky}
             uploadImages={imageProps && imageProps.upload}
