+++ packages/react-tinacms-editor/src/context/utils/plugin.ts
@@ -1,0 +1,9 @@
+import { Plugin } from '@tinacms/core'
+
+function byType(__type: string) {
+  return (plugin: Plugin) => plugin.__type === __type
+}
+
+export function findPlugins<P extends Plugin>(type: string, plugins: Plugin[]) {
+  return plugins.filter(byType(type)) as P[]
+}
