+++ packages/react-tinacms-editor/src/plugins/Link/Popups/Form.tsx
@@ -32,7 +32,7 @@ import {
   unmountLinkForm,
   updateLinkBeingEdited,
 } from '../commands'
-import { linkPluginKey } from '../index'
+import { linkPluginKey } from '../plugin'
 import { InnerForm } from './InnerForm'
 
 const width = 240
